+++ notification-settings/lib/generators/templates/install/initializer.rb
@@ -1,8 +1,11 @@
 # frozen_string_literal: true
 
 NotificationSettings.configure do |config|
-  # Choose your default notification category. Takes a string.
-  # config.default_category = 'notification'
+  # An array of all notification categories. Takes an array of symbols.
+  # config.categories = [:notification]
+
+  # Choose your default notification category. Takes a symbol.
+  # config.default_category = :notification
 
   ### STATUS ###
 
@@ -15,12 +18,21 @@ NotificationSettings.configure do |config|
   # config.offline_after = 3.hours
 
   # Stringified datetime attribute name of `object` that defines the time of the
-  # last activity. Takes a string.
-  # config.last_seen = 'last_seen'
+  # last activity. Takes a symbol.
+  # config.last_seen = :last_seen
+
+  # Array of all possible statuses. Takes an array of strings.
+  # config.statuses = [
+  #   'online',
+  #   'idle',
+  #   'offline',
+  #   'do not notify',
+  #   'do not disturb'
+  # ]
 
   # Array of possible statuses that will prevent creating notifications for a
   # target. Takes an array of strings.
-  # config.do_not_notify_statuses = []
+  # config.do_not_notify_statuses = ['do not notify']
 
   # Array of possible statuses that will prevent pushing notifications of a
   # target. Takes an array of strings.