+++ notification-settings/lib/notification_settings/preferences_form.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require 'active_model'
+
+module NotificationSettings
+  class PreferencesForm
+    include ActiveModel::Model
+
+    attribute :enabled, :boolean, default: true
+  end
+end