+++ notification-settings/spec/lib/notification_settings/notification_lib_spec.rb
@@ -5,13 +5,24 @@ require_relative '../../rails_helper'
 RSpec.describe NotificationSettings::NotificationLib do
   let(:notification) { build_stubbed :notification, category: :my_category }
   let(:user)         { notification.target }
-  let(:setting)      { user.notification_setting }
+  let(:settings)     { user.settings }
+
+  context 'with persisted user' do
+    let(:user) { create :user }
+
+    it 'persists settings' do
+      user.settings.enabled = true
+      user.save
+
+      expect(user.reload.settings.enabled?).to eq true
+    end
+  end
 
   it 'requires a target' do
     notification = Notification.new
     expect(notification.valid?).to eq false
     expect(notification.errors[:target])
-      .to eq ['must exist', "can't be blank"]
+      .to eq ['must exist']
     expect(notification.save).to eq false
     expect(notification.persisted?).to eq false
 
@@ -20,73 +31,64 @@ RSpec.describe NotificationSettings::NotificationLib do
     expect(notification.persisted?).to eq true
   end
 
-  describe ':enabled' do
-    it 'is considered enabled by default' do
-      expect(setting[:enabled]).to eq nil
+  describe 'global settings' do
+    it 'is enabled by default' do
+      expect(settings.enabled).to eq nil
       expect(notification.creation_allowed?).to eq true
       expect(user.notify.persisted?).to eq true
     end
 
-    it 'when disabled' do
-      setting.settings[:enabled] = false
+    it 'can be disabled' do
+      settings.enabled = false
       expect(notification.creation_allowed?).to eq false
       expect(user.notify.persisted?).to eq false
     end
-
-    describe 'category_settings' do
-      pending
-    end
   end
 
-  describe 'global (:delivery_method_enabled) settings' do
-    it 'is considered enabled by default' do
-      expect(setting[:delivery_method_enabled]).to eq nil
+  describe 'global pusher settings' do
+    it 'is enabled by default' do
+      expect(settings.delivery_methods_.enabled).to eq nil
 
       expect(notification.delivery_allowed?(:null)).to eq true
     end
 
-    it 'when disabled' do
-      setting.settings[:delivery_method_enabled] = false
+    it 'can be disabled' do
+      settings.delivery_methods!.enabled = false
 
-      expect(notification.delivery_allowed?(:null)).to eq false
-    end
-
-    it 'category_setting disabled' do
-      setting.category_settings[:my_category] = {}
-      setting.category_settings[:my_category][:delivery_method_enabled] = false
-
-      expect(notification.category_setting)
-        .to eq(delivery_method_enabled: false)
       expect(notification.delivery_allowed?(:null)).to eq false
     end
   end
 
-  describe 'pusher-specific settings' do
-    it 'is considered enabled by default' do
-      expect(setting[:null]).to eq nil
+  describe 'category-specific settings' do
+    it 'is enabled by default'
+
+    it 'can be disabled'
+
+    it 'can disable all delivery methods' do
+      settings.categories!.my_category!.delivery_methods!.enabled = false
+
+      expect(notification.delivery_allowed?(:null)).to eq false
+    end
+  end
+
+  describe 'delivery-method-specific settings' do
+    it 'is enabled by default' do
+      expect(settings.delivery_methods_.null).to eq nil
 
       expect(notification.delivery_allowed?(:null)).to eq true
       expect(user.notify.persisted?).to eq true
     end
 
-    it 'when disabled' do
-      setting.settings[:null] = false
-
-      expect(notification.delivery_allowed?(:null)).to eq false
-    end
-
-    it 'category_setting disabled' do
-      setting.category_settings[:my_category] = {}
-      setting.category_settings[:my_category][:null] = false
+    it 'can be disabled' do
+      settings.delivery_methods!.null = false
 
       expect(notification.delivery_allowed?(:null)).to eq false
     end
 
     it 'the enabled one still pushes with one pusher enabled ' \
        'and one disabled' do
-      setting.category_settings[:my_category] = {}
-      setting.category_settings[:my_category][:null] = false
-      setting.category_settings[:my_category][:some_pusher] = true
+      settings.categories!.my_category!.delivery_methods!.null = false
+      settings.categories!.my_category!.delivery_methods!.some_pusher = true
 
       expect(notification.delivery_allowed?(:null)).to                 eq false
       expect(notification.delivery_allowed?(:some_pusher)).to          eq true