+++ notification-settings/spec/support/rails_app/db/migrate/20190509202305_create_users.rb
@@ -4,6 +4,9 @@ class CreateUsers < ActiveRecord::Migration[5.2]
   def change
     create_table :users do |t|
       t.string :name
+
+      t.string :status
+      t.string :settings
     end
   end
 end